--- src/equicordplugins/instantScreenshare/instantScreenshare/index.tsx
@@ -1,184 +1,0 @@
-/*
- * Vencord, a Discord client mod
- * Copyright (c) 2025 Vendicated and contributors
- * SPDX-License-Identifier: GPL-3.0-or-later
- */
-
-import { getUserSettingLazy } from "@api/UserSettings";
-import { HeadingSecondary } from "@components/Heading";
-import { Paragraph } from "@components/Paragraph";
-import { Devs, EquicordDevs } from "@utils/constants";
-import definePlugin from "@utils/types";
-import { VoiceState } from "@vencord/discord-types";
-import { findByCodeLazy, findStoreLazy } from "@webpack";
-import { ChannelStore, FluxDispatcher, MediaEngineStore, PermissionsBits, PermissionStore, SelectedChannelStore, showToast, Toasts, UserStore, VoiceActions, VoiceStateStore } from "@webpack/common";
-
-import { getCurrentMedia, matchesKeybind, settings } from "./utils";
-
-let hasStreamed;
-const startStream = findByCodeLazy('type:"STREAM_START"');
-const StreamPreviewSettings = getUserSettingLazy("voiceAndVideo", "disableStreamPreviews")!;
-const ApplicationStreamingSettingsStore = findStoreLazy("ApplicationStreamingSettingsStore");
-
-async function autoStartStream() {
-    const selected = SelectedChannelStore.getVoiceChannelId();
-    if (!selected) return;
-
-    const channel = ChannelStore.getChannel(selected);
-    const isGuildChannel = !channel.isDM() && !channel.isGroupDM();
-
-    if (channel.type === 13 || isGuildChannel && !PermissionStore.can(PermissionsBits.STREAM, channel)) return;
-
-    if (settings.store.autoDeafen && !MediaEngineStore.isSelfDeaf()) {
-        VoiceActions.toggleSelfDeaf();
-    } else if (settings.store.autoMute && !MediaEngineStore.isSelfMute()) {
-        VoiceActions.toggleSelfMute();
-    }
-
-    const streamMedia = await getCurrentMedia();
-    const preview = StreamPreviewSettings.getSetting();
-    const { soundshareEnabled } = ApplicationStreamingSettingsStore.getState();
-    let sourceId = streamMedia.id;
-    if (streamMedia.type === "video_device") sourceId = `camera:${streamMedia.id}`;
-
-    startStream(channel.guild_id ?? null, selected, {
-        "pid": null,
-        "sourceId": sourceId,
-        "sourceName": streamMedia.name,
-        "audioSourceId": streamMedia.name,
-        "sound": soundshareEnabled,
-        "previewDisabled": preview
-    });
-}
-
-export default definePlugin({
-    name: "InstantScreenshare",
-    description: "Instantly screenshare when joining a voice channel with support for desktop sources, windows, and video input devices (cameras, capture cards)",
-    authors: [Devs.HAHALOSAH, Devs.thororen, EquicordDevs.mart],
-    dependencies: ["EquicordToolbox"],
-    getCurrentMedia,
-    settings,
-
-    settingsAboutComponent: () => (
-        <>
-            <HeadingSecondary>For Linux</HeadingSecondary>
-            <Paragraph>
-                For Wayland it only pops up the screenshare select
-                <br />
-                For X11 it may or may not work :shrug:
-            </Paragraph>
-            <br />
-            <HeadingSecondary>Video Devices</HeadingSecondary>
-            <Paragraph>
-                Supports cameras and capture cards (like Elgato HD60X) when enabled in settings
-            </Paragraph>
-            <br />
-            <HeadingSecondary>Regarding Sound & Preview Settings</HeadingSecondary>
-            <Paragraph>
-                We use the settings set and used by discord to decide if stream preview and sound should be enabled or not
-            </Paragraph>
-        </>
-    ),
-
-    start() {
-        window.addEventListener("keydown", onKeyDown, true);
-    },
-
-    stop() {
-        window.removeEventListener("keydown", onKeyDown, true);
-    },
-
-    flux: {
-        async VOICE_STATE_UPDATES({ voiceStates }: { voiceStates: VoiceState[]; }) {
-            if (!settings.store.toolboxManagement) return;
-            if (!settings.store.autoOnJoin) return;
-            const myId = UserStore.getCurrentUser().id;
-            for (const state of voiceStates) {
-                const { userId, channelId } = state;
-                if (userId !== myId) continue;
-
-                if (channelId && !hasStreamed) {
-                    hasStreamed = true;
-                    await autoStartStream();
-                }
-
-                if (!channelId) {
-                    hasStreamed = false;
-                }
-
-                break;
-            }
-        }
-    },
-
-    toolboxActions: {
-        "Instant Screenshare"() {
-            settings.store.toolboxManagement = !settings.store.toolboxManagement;
-            showToast(`Instant Screenshare ${settings.store.toolboxManagement ? "Enabled" : "Disabled"}`, Toasts.Type.SUCCESS);
-        }
-    }
-});
-
-function onKeyDown(e: KeyboardEvent) {
-    const kb = settings.store.keybind;
-    if (!kb) return;
-    if (matchesKeybind(e, kb)) {
-        const target = e.target as HTMLElement | null;
-        if (target && (target.tagName === "INPUT" || target.tagName === "TEXTAREA" || target.isContentEditable)) return;
-        e.preventDefault();
-        e.stopPropagation();
-        toggleScreenShare();
-    }
-}
-
-async function toggleScreenShare() {
-    try {
-        const userId = UserStore.getCurrentUser().id;
-        const voiceState = VoiceStateStore.getVoiceStateForUser(userId);
-        const channelId = SelectedChannelStore.getVoiceChannelId?.() ?? voiceState?.channelId;
-        if (!channelId) {
-            showToast("Join a voice channel first", Toasts.Type.FAILURE);
-            return;
-        }
-        const channel = ChannelStore.getChannel(channelId);
-        if (!channel) {
-            showToast("Unable to resolve channel", Toasts.Type.FAILURE);
-            return;
-        }
-
-        if (voiceState?.selfStream) {
-            const streamKey = channel.guild_id
-                ? `guild:${channel.guild_id}:${channelId}:${userId}`
-                : `call:${channelId}:${userId}`;
-            FluxDispatcher.dispatch({ type: "STREAM_STOP", streamKey, appContext: "APP" });
-            showToast("Stopped screenshare", Toasts.Type.SUCCESS);
-            return;
-        }
-
-        const media = MediaEngineStore.getMediaEngine();
-        const sources = await findDesktopSources(media);
-        const screen = sources[0];
-        if (!screen) {
-            showToast("No desktop sources found", Toasts.Type.FAILURE);
-            return;
-        }
-
-        startStream(channel.guild_id ?? null, channelId, {
-            pid: null,
-            sourceId: screen.id,
-            sourceName: screen.name,
-            audioSourceId: screen.name,
-            sound: true,
-            previewDisabled: false
-        });
-        showToast(`Sharing: ${screen.name}`, Toasts.Type.SUCCESS);
-    } catch (err) {
-        console.error("InstantScreenshare hotkey toggle failed", err);
-        showToast("Failed to toggle screenshare (see console)", Toasts.Type.FAILURE);
-    }
-}
-
-async function findDesktopSources(media) {
-    const sources = await findByCodeLazy("desktop sources")(media, ["screen"], null) ?? [];
-    return sources;
-}
